# Conflicts:
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/find/dailyperform/DailyRecordDto.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/actualworkinghours/AttendanceTimeOfDailyPerformance.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyprocess/calc/AdTimeAndAnyItemAdUpServiceImpl.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyprocess/calc/CalculateDailyRecordServiceCenterImpl.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyprocess/calc/CalculateDailyRecordServiceImpl.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyprocess/calc/CommonCompanySettingForCalcImpl.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyprocess/calc/DailyCalculationEmployeeServiceImpl.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyprocess/calc/FactoryManagePerPersonDailySetImpl.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/require/RecordDomRequireService.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/dailyattdcal/dailycalprocess/calculation/other/HolidayWorkFrameTimeSheetForCalc.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/dailyattdcal/dailycalprocess/calculation/other/HolidayWorkTimeSheet.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/dailyattdcal/dailycalprocess/calculation/other/ManagePerPersonDailySet.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/dailyattdcal/dailycalprocess/calculation/other/OutsideWorkTimeSheet.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattdcal/dailyattendance/converter/util/ItemConst.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattdcal/dailyattendance/premiumtime/PremiumTimeOfDailyPerformance.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattdcal/dailyattendance/worktime/ActualWorkingTimeOfDaily.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattdcal/dailyattendance/worktime/AttendanceTimeOfDailyAttendance.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattdcal/dailycalprocess/calculation/timezone/CalculationRangeOfOneDay.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattdcal/dailycalprocess/calculation/timezone/CalculationTimeSheet.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattdcal/dailycalprocess/calculation/timezone/outsideworktime/OverTimeFrameTimeSheetForCalc.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattdcal/dailycalprocess/calculation/timezone/outsideworktime/OverTimeSheet.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattdcal/ortherpackage/classfunction/ManageReGetClass.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattdcal/ortherpackage/classfunction/PersonnelCostSettingImport.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/scherec/dailyattendanceitem/enums/PrimitiveValueOfAttendanceItem.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/standardtime/KmkmtAgeementYearSetting.java
@@ -24,10 +24,10 @@ public class KmkmtAgeementYearSetting extends UkJpaEntity implements Serializabl
 	public KmkmtAgeementYearSettingPK kmkmtAgeementYearSettingPK;
 	
 	@Column(name ="ERROR_YEARLY")
-	public BigDecimal errorOneYear;
+	public int errorOneYear;
 	
 	@Column(name ="ALARM_YEARLY")
-	public BigDecimal alarmOneYear;
+	public int alarmOneYear;
 	
 	@Override
 	protected Object getKey() {
